Modularising and reorganising the fast explainer Swagger subsystem
